% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-colormap.R
\name{as_ieegio_colortable}
\alias{as_ieegio_colortable}
\title{Convert to ieegio color table}
\usage{
as_ieegio_colortable(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{additional arguments passed to methods}
}
\value{
An \code{ieegio_colortable} object.
}
\description{
Convert a data frame or existing label table to an \code{ieegio_colortable}
object.
}
\examples{
ct <- as_ieegio_colortable(data.frame(
  Key = 1:3,
  R   = c(255L, 0L, 0L),
  G   = c(0L, 255L, 0L),
  B   = c(0L, 0L, 255L),
  A   = c(255L, 255L, 255L)
))
print(ct)
}
