% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti.R
\name{as_nifti_header}
\alias{as_nifti_header}
\title{Internal method to extract header information from a 'NIfTI' file}
\usage{
as_nifti_header(x)
}
\arguments{
\item{x}{file path or an R object}
}
\value{
A list containing the file header information
}
\description{
Internal method to extract header information from a 'NIfTI' file
}
