% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{io-tt}
\alias{io-tt}
\alias{io_read_tt}
\title{Read \code{'TT'} streamline file}
\usage{
io_read_tt(file)
}
\arguments{
\item{file}{path to the streamline file}
}
\value{
An \code{\link{imaging-streamlines}} instance.
}
\description{
Writer is not implemented yet. Please save as a \code{'TCK'} file.
}
\examples{


# This example uses sample data, run
# `ieegio_sample_data("streamlines/CNVII_R.trk")` to download

if( ieegio_sample_data("streamlines/CNVII_R.tt.gz", test = TRUE) ) {

  path <- ieegio_sample_data("streamlines/CNVII_R.tt.gz")

  # read
  x <- io_read_tt(path)

  plot(x)

}

}
