% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-volume.R
\name{merge.ieegio_volume}
\alias{merge.ieegio_volume}
\title{Merge \code{'ieegio'} volumes}
\usage{
\method{merge}{ieegio_volume}(x, y, ..., thresholds = 0, reshape = dim(x), na_fill = NA)
}
\arguments{
\item{x}{base image to be merged}

\item{y, ...}{images to be merged into \code{x}}

\item{thresholds}{numerical threshold for \code{y,...}, can be length of one
or more, if images to overlay is more than one. The image values lower
than the threshold will be trimmed out}

\item{reshape}{output shape, default is the dimension of \code{x}; if
changed, then the underlay will be sampled into the new shape}

\item{na_fill}{how to handle missing values; default is \code{NA}; for
compatibility, you might want to set to 0}
}
\value{
Merged volume with dimension \code{reshape}.
}
\description{
Merge volume data into base image. The images must be static 3-dimensional
volume data. Currently time-series or 4-dimensional data is not supported.
}
\examples{



base_array <- array(0, c(15, 15, 15))
base_array[4:6, 4:6, 4:6] <- 1

# generate a 15x15x15 mask with 1mm spacing
vox2ras1 <- diag(1, 4)
vox2ras1[1:3, 4] <- -5
x <- as_ieegio_volume(base_array, vox2ras = vox2ras1)
plot(x,
     zoom = 10,
     position = c(0, 0, 0),
     pixel_width = 0.5)

# 15x15x15 mask with 0.5mmx1mmx1mm spacing
vox2ras2 <- diag(c(2, 1, 1, 1), 4)
vox2ras2[1:3, 4] <- c(-3,-4, -1)
y <- as_ieegio_volume(base_array, vox2ras = vox2ras2)
plot(y,
     zoom = 10,
     position = c(0, 0, 0),
     pixel_width = 0.5)

# merge y into x and up-sample to 30x30x30 mask with 0.5mm spacing
z <- merge(x, y, reshape = c(30, 30, 30))
plot(
  z,
  zoom = 10,
  position = c(0, 0, 0),
  pixel_width = 0.5
)

}
