% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti.R
\name{nii}
\alias{nii}
\alias{read_nii}
\alias{write_nii}
\alias{write_nii.ieegio_nifti}
\alias{write_nii.ants.core.ants_image.ANTsImage}
\alias{write_nii.niftiImage}
\alias{write_nii.nifti}
\alias{write_nii.ieegio_mgh}
\alias{write_nii.array}
\title{Read and write 'NIfTI' imaging data}
\usage{
read_nii(file, format = c("oro", "rnifti", "ants"), header_only = FALSE, ...)

write_nii(x, con, ...)

\method{write_nii}{ieegio_nifti}(x, con, ...)

\method{write_nii}{ants.core.ants_image.ANTsImage}(x, con, ...)

\method{write_nii}{niftiImage}(x, con, ...)

\method{write_nii}{nifti}(x, con, ...)

\method{write_nii}{ieegio_mgh}(x, con, ...)

\method{write_nii}{array}(x, con, vox2ras = NULL, ...)
}
\arguments{
\item{file}{file path to read volume data}

\item{format}{header format; choices are \code{'oro'} (using
\code{\link[oro.nifti]{readNIfTI}}), \code{'rnifti'} (using
\code{\link[RNifti]{readNifti}}), and \code{'ants'} (using
\code{\link[rpyANTs]{as_ANTsImage}}).}

\item{header_only}{whether to read header data only;
default is \code{FALSE}}

\item{...}{passed to other methods, mostly
\code{\link[freesurferformats]{write.fs.mgh}}}

\item{x}{volume data (such as 'NIfTI' image, array, or 'MGH')
to be saved}

\item{con}{file path to store image}

\item{vox2ras}{a \code{4x4} transform matrix from voxel indexing (column,
row, slice) to scanner (often 'T1-weighted' image) 'RAS'
(right-anterior-superior) coordinate}
}
\value{
\code{read_nii} returns an \code{ieegio_volume} object
}
\description{
Read and write 'NIfTI' imaging data
}
\examples{

if(ieegio_sample_data(
  file = "brain.demosubject.nii.gz",
  test = TRUE)) {

  file <- ieegio_sample_data("brain.demosubject.nii.gz")

  # ---- basic read -----------------------------------------------
  vol <- read_nii(file)

  # voxel to scanner RAS
  vol$transforms$vox2ras

  # to freesurfer surface
  vol$transforms$vox2ras_tkr

  # to FSL
  vol$transforms$vox2fsl

  image(vol$data[,,128], asp = 1, axes = FALSE)


  # ---- using other methods --------------------------------------
  # default
  vol <- read_nii(file, format = "oro")
  vol$header

  # lazy-load nifti
  vol2 <- read_nii(file, format = "rnifti")
  vol2$header

  # Using ANTsPyx
  vol3 <- read_nii(file, format = "ants")
  vol3$header

  # ---- write --------------------------------------------------------
  f <- tempfile(fileext = ".nii.gz")
  write_nii(vol, f)
  write_nii(vol2, f)
  write_nii(vol3, f)

  # other ways
  write_nii(vol$header, f)
  write_nii(vol2$header, f)
  write_nii(vol3$header, f)

  # try to save MGH/MGZ data to NIfTI
  file2 <- ieegio_sample_data("brain.demosubject.mgz")
  mgz <- read_mgz(file2)
  write_nii(mgz, f)

  # clean up
  unlink(f)

}


}
\seealso{
\code{\link{mgz}}
}
