% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py-setup.R
\name{pynwb_module}
\alias{pynwb_module}
\alias{install_pynwb}
\title{Install \code{'NWB'} via \code{'pynwb'}}
\usage{
install_pynwb(python_ver = "auto", verbose = TRUE)

pynwb_module(force = FALSE, error_if_missing = TRUE)
}
\arguments{
\item{python_ver}{'Python' version, see \code{\link[rpymat]{configure_conda}};
default is \code{"auto"}, which is suggested}

\item{verbose}{whether to print the installation messages}

\item{force}{whether to force-reload the module}

\item{error_if_missing}{whether to raise errors when the module fails to
load; default is true}
}
\value{
A 'Python' module \code{pynwb}.
}
\description{
Install \code{'NWB'} via \code{'pynwb'}
}
