% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-docs.R, R/freesurfer.R,
%   R/generics-surface.R, R/gifti.R
\name{read_surface}
\alias{read_surface}
\alias{io_read_fs}
\alias{io_read_gii}
\title{Read surface files}
\usage{
io_read_fs(
  file,
  type = c("geometry", "annotations", "measurements"),
  format = "auto",
  name = basename(file),
  ...
)

read_surface(file, format = "auto", type = NULL, ...)

io_read_gii(file)
}
\arguments{
\item{file}{path the file}

\item{type}{type of the data; ignored if the file format is 'GIfTI'. For
'FreeSurfer' files, supported types are
\describe{
\item{\code{'geometry'}}{contains positions of mesh vertex nodes and face indices;}
\item{\code{'annotations'}}{annotation file (usually with file extension \code{'annot'}) containing a color look-up table and an array of color keys. These files are used to display discrete values on the surface such as brain atlas;}
\item{\code{'measurements'}}{measurement file such as \code{'sulc'} and \code{'curv'} files, containing numerical values (often with continuous domain) for each vertex node}
}}

\item{format}{format of the file, see 'Arguments' section in
\code{\link[freesurferformats]{read.fs.surface}} (when file type is
\code{'geometry'}) and \code{\link[freesurferformats]{read.fs.curv}}
(when file type is \code{'measurements'})}

\item{name}{name of the data; default is the file name}

\item{...}{for \code{read_surface}, the arguments will be passed to
\code{io_read_fs} if the file is a 'FreeSurfer' file.}
}
\value{
A surface object container
}
\description{
Supports reading surface geometry, annotation, measurement, and
time-series data.
Please use the high-level function \code{read_surface}, which calls
other low-level functions internally.
}
